refactor: trip edit folder structure
