+++ src/features/trip/edit/components/Tabs/Photo.tsx
