+++ src/features/trip/list/components/AllTrips.tsx
@@ -2,9 +2,9 @@ import AddIcon from '@mui/icons-material/Add';
 import { Box, CircularProgress, Link, Stack, Typography } from '@mui/material';
 
 import { AppRoutes } from '@config/routes';
+import { useGetTripsQuery } from '@features/trip/store/tripsApi';
 import AppButton from '@features/ui/AppButton';
 
-import { useGetTripsQuery } from '../store/tripsApi';
 import NoTrips from './NoTrips';
 import TripsList from './TripsList';
