+++ src/pages/account/trips/trips.tsx
@@ -1,4 +1,4 @@
-import AllTrips from '@features/trip/list/AllTrips';
+import AllTrips from '@features/trip/list/components/AllTrips';
 
 export default function TripsPage() {
   return <AllTrips />;
